--- tests/conftest.py
@@ -1,81 +1,0 @@
-import polars as pl
-import pytest
-
-
-@pytest.fixture
-def frame():
-    """
-    Make a mock data frame to test model building.
-    """
-    frame = pl.DataFrame(
-        {
-            "geography": ["USA"] * 20,
-            "time_end": [
-                "2018-07-01",
-                "2018-08-01",
-                "2018-09-01",
-                "2018-10-01",
-                "2018-11-01",
-                "2018-12-01",
-                "2019-01-01",
-                "2019-02-01",
-                "2019-03-01",
-                "2019-04-01",
-                "2019-07-01",
-                "2019-08-01",
-                "2019-09-01",
-                "2019-10-01",
-                "2019-11-01",
-                "2019-12-01",
-                "2020-01-01",
-                "2020-02-01",
-                "2020-03-01",
-                "2020-04-01",
-            ],
-            "lci": [
-                0.0005,
-                0.0015,
-                0.0025,
-                0.0035,
-                0.0045,
-                0.0055,
-                0.0065,
-                0.0075,
-                0.0085,
-                0.0095,
-            ]
-            * 2,
-            "uci": [
-                0.0015,
-                0.0025,
-                0.0035,
-                0.0045,
-                0.0055,
-                0.0065,
-                0.0075,
-                0.0085,
-                0.0095,
-                0.0105,
-            ]
-            * 2,
-            "sample_size": [1000, 1000, 2000, 3000, 4000, 4000, 4000, 5000, 4000, 5000]
-            * 2,
-            "estimate": [
-                0.001,
-                0.002,
-                0.003,
-                0.004,
-                0.005,
-                0.006,
-                0.007,
-                0.008,
-                0.009,
-                0.01,
-            ]
-            * 2,
-            "season": ["2018/2019"] * 10 + ["2019/2020"] * 10,
-        },
-        schema_overrides={"time_end": pl.Date},
-    )
-
-    return frame
